requirements.txt
numba==0.55.2
numpy==1.22.4
opencv-python==4.6.0.66
pillow==9.1.1
pydantic==1.9.1
